Trying to fix PropertyReferenceException
Shortened some methods to make them easier to understand, and added getEmployeeID methods to Compensations and ReportingStructures so that there is easy access to employee IDs directly from the structures themselves. Currently still receiving a PropertyReferenceException that occurs when trying to run Gradlew bootRun. Working to resolve that before moving on to tests.