Changing ReportingStrcuture thought process
Given the wording of how ReportingStructure is, I am no longer treating it like it should be something saved into a database, and instead going to treat it like a read which calls upon a calculation.

This means that when you call a ReportingStructure read (REST GET command), it goes through and starts doing the calculations all on it's own from there. No database required.