Tests now pass for ReportingStructure
The test for Reporting Structure is passing!

Had to rewrite how reports were being calculated, changing the method to tallyReports and using a recursive method to get the total number of reports. This tally, while stored, does get recalculated every time you request the number of reports for an employee.